Rename file to `bootstrap_service`
